# football-stats-go-api/docs/swagger.yaml
basePath: /api
definitions:
  example_hello_internal_domain.DBXGStatFixture:
    properties:
      away_score:
        type: integer
      away_shots:
        items:
          $ref: '#/definitions/example_hello_internal_domain.DBXGStatShot'
        type: array
      away_team:
        type: string
      away_xg:
        type: number
      date:
        type: string
      gameweek:
        type: integer
      home_score:
        type: integer
      home_shots:
        items:
          $ref: '#/definitions/example_hello_internal_domain.DBXGStatShot'
        type: array
      home_team:
        type: string
      home_xg:
        type: number
      id:
        type: integer
    type: object
  example_hello_internal_domain.DBXGStatShot:
    properties:
      is_goal:
        type: boolean
      minute:
        type: integer
      player_name:
        type: string
      shot_type:
        type: string
      x:
        type: number
      xg:
        type: number
      "y":
        type: number
    type: object
  internal_api.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  internal_api.ScrapeRequest:
    properties:
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@footballstats.com
    name: API Support
  description: A lightweight API for scraping football data from websites
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Football Stats Scraper API
  version: "1.0"
paths:
  /scrape/xgstats:
    post:
      consumes:
      - application/json
      description: Scrape xG statistics and shot map data from xgstat.com
      parameters:
      - description: Scrape request with xgstat.com URL
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.ScrapeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Scraped xG statistics and shot map data
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/example_hello_internal_domain.DBXGStatFixture'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Scrape xG shot map data
      tags:
      - scraper
  /xgstats:
    get:
      consumes:
      - application/json
      description: Retrieve saved xG statistics and shot map data from the database
      parameters:
      - description: Fixture ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved xG statistics
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/example_hello_internal_domain.DBXGStatFixture'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "404":
          description: Fixture not found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get xG statistics by fixture ID
      tags:
      - scraper
schemes:
- http
- https
swagger: "2.0"